09: dev_tools

+++ Source/core/fetch/ResourceTest.cpp
@@ -6,9 +6,9 @@
 #include "core/fetch/Resource.h"
 
 #include "core/fetch/ResourcePtr.h"
-#include "core/testing/URLTestHelpers.h"
 #include "platform/network/ResourceRequest.h"
 #include "platform/network/ResourceResponse.h"
+#include "platform/testing/URLTestHelpers.h"
 #include "public/platform/Platform.h"
 #include "wtf/Vector.h"
 
@@ -26,7 +26,7 @@ public:
     ~MockPlatform() override { }
 
     // From blink::Platform:
-    void cacheMetadata(const WebURL& url, double, const char*, size_t) override
+    void cacheMetadata(const WebURL& url, int64, const char*, size_t) override
     {
         m_cachedURLs.append(url);
     }
